Extract useStatisticsData TODO: fix statistics not being fetched when in page > 1
